first attempt at using the package
