+++ doc/.gitignore
@@ -1,0 +1,3 @@
+algo_doc/
+analysis/
+old_src/
